feat: extract NotesClient component
