+++ src/app/notes/page.tsx
@@ -5,13 +5,10 @@ import { redirect } from 'next/navigation'
 import { createNote } from '@/app/actions'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
-import { NotesList, Note } from './NotesList'
+import { Note } from './NotesList'
 import { countOpenTasks } from '@/lib/taskparse'
 import { NavButton } from '@/components/NavButton'
-import ViewSelector from '@/components/ViewSelector'
-import FilterBar, { NoteFilters } from '@/components/notes/FilterBar'
-import { Filter, LayoutPanelTop, LayoutGrid, List } from 'lucide-react'
-import { useState, useMemo } from 'react'
+import { NotesClient } from './NotesClient'
 
 export default async function NotesPage() {
   const supabase = await supabaseServer()
@@ -23,7 +20,7 @@ export default async function NotesPage() {
     .select('id,title,updated_at,body')
     .order('updated_at', { ascending: false })
 
-  const enriched = (notes ?? []).map(n => ({
+  const enriched: Note[] = (notes ?? []).map(n => ({
     id: n.id,
     title: n.title,
     updated_at: n.updated_at,
@@ -49,50 +46,6 @@ export default async function NotesPage() {
       </div>
 
       <NotesClient notes={enriched} />
-    </div>
-  )
-}
-
-export function NotesClient({ notes }: { notes: Note[] }) {
-  'use client'
-  const [showFilters, setShowFilters] = useState(false)
-  const [filters, setFilters] = useState<NoteFilters>({ sort: 'newest' })
-
-  const filtered = useMemo(() => {
-    let res = [...notes]
-    if (filters.search) {
-      const s = filters.search.toLowerCase()
-      res = res.filter(n => n.title?.toLowerCase().includes(s))
-    }
-    res.sort((a, b) => {
-      const diff = new Date(a.updated_at).getTime() - new Date(b.updated_at).getTime()
-      return filters.sort === 'oldest' ? diff : -diff
-    })
-    return res
-  }, [notes, filters])
-
-  return (
-    <div className="space-y-3">
-      <div className="flex items-center gap-2">
-        <ViewSelector
-          defaultValue="card"
-          options={[
-            { value: 'card', label: 'Card', icon: LayoutPanelTop },
-            { value: 'grid', label: 'Grid', icon: LayoutGrid },
-            { value: 'list', label: 'List', icon: List },
-          ]}
-        />
-        <button
-          type="button"
-          aria-label="Toggle filters"
-          onClick={() => setShowFilters(s => !s)}
-          className="rounded-md border border-input p-2 hover:bg-accent/50"
-        >
-          <Filter className="size-4" />
-        </button>
-      </div>
-      {showFilters && <FilterBar onChange={setFilters} />}
-      <NotesList notes={filtered} />
     </div>
   )
 }
